+++ dnsimple/service/registrar.py
@@ -1,4 +1,5 @@
 import json
+import warnings
 
 from dnsimple.response import Response
 from dnsimple.struct import DomainCheck, DomainPremiumPrice, DomainRegistration, DomainTransfer, DomainRenewal, \
@@ -34,7 +35,8 @@ class Registrar(object):
 
     def get_domain_premium_price(self, account_id, domain, options=None):
         """
-        Get the premium price for a domain.
+        DEPERECATED: Get the premium price for a domain.
+        Use get_domain_prices
 
         See https://developer.dnsimple.com/v2/registrar/#getDomainPremiumPrice
 
@@ -49,6 +51,9 @@ class Registrar(object):
         :return: dnsimple.Response
             The domain premium price requested
         """
+
+        warnings.warn("DEPRECATION WARNING: get_domain_premium_price is deprecated, use get_domain_prices instead.")
+
         if options is None:
             options = {}
         response = self.client.get(f'/{account_id}/registrar/domains/{domain}/premium_price', params=options)