+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 - CHANGED:
   - Feature: Added `registrar.get_domain_prices` to retrieve whether a domain is premium and the prices to register, transfer, and renew. (dnsimple/dnsimple-python#174)
+  - Deprecated: get_domain_premium_price, use get_domain_prices instead.
   - Dependency updates
 
 ## 2.1.2